Refactor highlighting to use our own method
The old design was simpler on our end but it had the disadvantage of
highlighting targets regardless of their scope.  This was confusing as
it would show matches outside the context of the given command.

Thanks to Tomasz Hołubowicz for bringing this matter to my attention
and for Jeremy Clark for recommending a look at the 'symbol-overlay'
package:
https://lists.sr.ht/~protesilaos/general-issues/%3C87y1q27400.fsf%40alternateved.com%3E