Enable empty string substitution
We could not do that with 'read-string' as the empty input is
automatically interpreted as the value of the DEFAULT-VALUE argument.
This is standard for most Emacs minibuffer interactions, though in our
case it makes sense to accept an empty string, as we may, e.g., want
to remove a prefix from a target.

Thanks to Kostas Andreadis for discussing this with me in issue 2 on
the GitHub mirror as well as pull request 3:

- <https://github.com/protesilaos/substitute/issues/2>
- <https://github.com/protesilaos/substitute/pull/3>